Remove hard-coded player one name
This is not a really a refactoring step, rather a small and cheap
improvement.

Player one name was harcoded, making it totally useless  storing
their name in `@player1Name` (`@player2Name` is still not  used
anywhere, but I'm going to make another related improvement
later).

Alternatively, instead of improving the codebase the alternative
would have been to remove some useless code (player name variables)
but I think that this little improvement is what makes more sense
for the customer, if not right now, it will do it in the future,
as having hard-coded names is not a good idea in the long term.