Use hash to store players data
By using a ruby hash we can simplify the number of instance variables
used in the class to a single one. This has the extra benefit of
simplifying the rest of the codebase as well. The new code is possible
thanks to the fact that ruby hashes are positionally ordered since
version 1.9